MIDAPSim/requirements.txt
future >= 0.16.0
Pillow >= 5.2.0
numpy >= 1.15.1
torch >= 1.0.0
scikit-image >= 0.14.3 
pytest >= 5.4.1
pytest-parallel
orderedset
pyyaml == 5.4.1
pandas
graphviz >= 0.19.1
matplotlib >= 3.5.1
python-box
pyqt5
